add instance and template classes
